dashboard/requirements.txt
streamlit
psycopg2-binary
pandas
seaborn
matplotlib
plotly